+++ e2e/cypress/support/commands.js
@@ -44,10 +44,10 @@ const getUrl = (componentId, fixtureName = '', config = {}) => {
 };
 
 const run = (app, fixtureName, plugins) => {
-  cy.visit(getUrl(app, fixtureName, plugins)).then(() => {
+  cy.visit(getUrl(app, fixtureName, plugins)).then(contentWindow => {
     disableTransitions();
     findEditorElement();
-    hideAllTooltips();
+    contentWindow.richContentHideTooltips = true;
   });
 };
 
@@ -79,10 +79,6 @@ Cypress.Commands.add('switchToEnglish', () => {
 
 function disableTransitions() {
   Cypress.$('head').append('<style> * {transition: none !important;}</style>');
-}
-
-function hideAllTooltips() {
-  cy.get('[data-id="tooltip"]', { timeout: 60000 }).invoke('hide'); //uses jquery to set display: none
 }
 
 function findEditorElement() {