+++ examples/main/src/Components/ExternalToolbar.jsx
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { FileInput, Tooltip, BUTTON_TYPES } from 'wix-rich-content-editor-common';
+import { FileInput, BUTTON_TYPES } from 'wix-rich-content-editor-common';
+import Tooltip from 'wix-rich-content-common/dist/lib/Tooltip.cjs.jsx';
+
 import styles from './ExternalToolbar.scss';
 
 class ExternalToolbar extends Component {
@@ -12,8 +14,9 @@ class ExternalToolbar extends Component {
     const { onClick, getIcon, dataHook, isDisabled, isActive, tooltip } = buttonProps;
     const Icon = getIcon();
     const style = isActive() ? { background: 'lightslategray' } : {};
+
     return (
-      <Tooltip content={tooltip} place="bottom" moveBy={{ y: -20 }}>
+      <Tooltip content={tooltip} place="bottom" tooltipOffset={{ y: -15 }}>
         <button disabled={isDisabled()} data-hook={dataHook} onClick={onClick} style={style}>
           <Icon />
         </button>
@@ -41,7 +44,7 @@ class ExternalToolbar extends Component {
         multiple={multiple}
         key={name}
       >
-        <Tooltip content={tooltip} place="bottom" moveBy={{ y: -20 }}>
+        <Tooltip content={tooltip} place="bottom" tooltipOffset={{ y: -15 }}>
           <Icon />
         </Tooltip>
       </FileInput>