+++ examples/storybook/stories/Components/InitialIntentToolbar.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { pick } from 'lodash';
-import { FileInput, Tooltip, TooltipHost, BUTTON_TYPES } from 'wix-rich-content-editor-common';
+import { FileInput, BUTTON_TYPES } from 'wix-rich-content-editor-common';
+import Tooltip from 'wix-rich-content-common/dist/lib/Tooltip.cjs.jsx';
 import PhotoCamera from 'wix-ui-icons-common/PhotoCamera';
 import VideoCamera from 'wix-ui-icons-common/VideoCamera';
 import styles from './InitialIntentToolbar.css';
@@ -71,7 +72,6 @@ class InitialIntentToolbar extends Component {
             return null;
           }
         )}
-        <TooltipHost />
       </div>
     );
   }