+++ packages/editor-common/web/src/Base/createBaseInsertPluginButton.jsx
@@ -155,7 +155,7 @@ export default ({
           theme={theme}
           tooltipText={buttonProps.tooltip}
           button={Button}
-          tooltipOffset={{ y: -10 }}
+          tooltipOffset={{ y: 0 }}
         />
       );
     }