+++ packages/editor-common/web/src/Components/InfoIcon.jsx
@@ -1,14 +1,13 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import generalstyles from '../../statics/styles/general.scss';
-import Tooltip from './Tooltip';
+import Tooltip from 'wix-rich-content-common/dist/lib/Tooltip.cjs.jsx';
 import Icon from '../Icons/InfoIcon';
 
 class InfoIcon extends Component {
   render() {
     const { tooltipText, iconStyles } = this.props;
     const style = iconStyles || generalstyles.infoIcon;
-
     return (
       <Tooltip content={tooltipText}>
         <Icon className={style} />