+++ packages/editor-common/web/src/Components/TextInput.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-
 import { mergeStyles } from 'wix-rich-content-common';
+import Tooltip from 'wix-rich-content-common/dist/lib/Tooltip.cjs.jsx';
 import { ErrorIcon } from '../Icons';
-import Tooltip from './Tooltip';
 import textInputStyles from '../../statics/styles/text-input.scss';
 import { omit } from 'lodash';
 
@@ -31,6 +30,7 @@ export default class TextInput extends React.Component {
     const { inputRef, error, theme, showTooltip, ...otherProps } = this.props;
     const inputProps = omit(otherProps, ['onChange']);
     const styles = mergeStyles({ styles: textInputStyles, theme });
+
     return (
       <div className={styles.textInput}>
         <input
@@ -43,13 +43,7 @@ export default class TextInput extends React.Component {
         />
         {error &&
           (showTooltip ? (
-            <Tooltip
-              shouldRebuildOnUpdate={() => !!error}
-              content={error}
-              theme={theme}
-              moveBy={{ y: 0 }}
-              type={'error'}
-            >
+            <Tooltip isError content={error}>
               <ErrorIcon className={styles.textInput_errorIcon} />
             </Tooltip>
           ) : (