+++ packages/editor/web/src/RichContentEditor/Toolbars/buttons/utils/createTextAlignmentButton.jsx
@@ -15,7 +15,6 @@ export default ({ alignment, Icon, tooltipTextKey }) =>
       isMobile: PropTypes.bool,
       t: PropTypes.func,
       tabIndex: PropTypes.number,
-      shouldRefreshTooltips: PropTypes.func,
     };
 
     isActive = () => this.props.alignment === alignment;
@@ -31,7 +30,7 @@ export default ({ alignment, Icon, tooltipTextKey }) =>
     };
 
     render() {
-      const { theme, isMobile, t, tabIndex, shouldRefreshTooltips } = this.props;
+      const { theme, isMobile, t, tabIndex } = this.props;
       const tooltipText = t(tooltipTextKey);
       const dataHookText = `textAlignmentButton_${alignment}`;
 
@@ -45,7 +44,6 @@ export default ({ alignment, Icon, tooltipTextKey }) =>
           tooltipText={tooltipText}
           dataHook={dataHookText}
           tabIndex={tabIndex}
-          shouldRefreshTooltips={shouldRefreshTooltips}
         />
       );
     }