+++ packages/editor/web/src/RichContentEditor/Toolbars/buttons/utils/createTextInlineStyleButton.jsx
@@ -47,6 +47,7 @@ export default ({ style, Icon, tooltipTextKey }) =>
           tooltipText={tooltipText}
           dataHook={dataHookText}
           tabIndex={tabIndex}
+          tooltipOffset={{ y: -20 }}
         />
       );
     }