+++ packages/plugin-link/web/src/toolbar/TextLinkButton.jsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import ReactTooltip from 'react-tooltip';
 import PropTypes from 'prop-types';
 import {
   hasLinksInSelection,
@@ -10,7 +9,6 @@ import {
 
 export default class TextLinkButton extends Component {
   showLinkPanel = () => {
-    ReactTooltip.hide();
     const {
       getEditorState,
       setEditorState,