+++ packages/plugin-social-polls/web/src/components/poll-option/PollGridOption.jsx
@@ -1,14 +1,10 @@
 import React from 'react';
 import cls from 'classnames';
-
-import { Tooltip } from 'wix-rich-content-editor-common';
-
+import Tooltip from 'wix-rich-content-common/dist/lib/Tooltip.cjs.jsx';
 import { RemoveIcon, NonEditableIcon, CheckIcon, LoaderIcon } from '../../assets/icons';
 import { TextField } from '../text-field';
 import { ImageUpload } from '../image-upload';
-
 import { PollOptionBase } from './PollOptionBase';
-
 import styles from './option.scss';
 
 export class PollGridOption extends PollOptionBase {